added a find method to the CircularBytes class. This will take care of bytes.find over the origin. Leads to simpler Dseq.find implementation.
Added a __bytes__ dunder method to Dseq.
Previously, bytes(Dseq) called __bytes__ on the Bio.Seq.Seq
returning basically the ._data property.
This is different from str(Dseq) that returns the watson strand
of an imaginary blunt version of the Dseq.
With this, bytes and str returns the same thing ecxept for encoding.

Explain the added "79.0" in mw method.

Added that Dseq.find is case sensitive. I think this is
the least surprising functionality. Debatable.

simplified the docstring for Dseq.find method

removed commented out code

reimplemented __contains__ in terms of Dseq.find

Highlighted that __eq__ is case insensitive. Debatable.

Added docstring to shifted method

simplified the looped by testing if self+self gives a TypeError.
This puts the logics in the __add__